+++ test/fixtures/export-error/error-obj-method.js
@@ -3,12 +3,12 @@
 This tests an error because the definitions of fn collide.
  */
 
-export function fn() {
+export function a() {
   return "X"
 }
 
 export default {
-  fn() {
+  a() { // conflicting definition from the named export a
     return "Y"
   }
 }