+++ test/fixtures/export-mixed/export-mixed-anon-func.js
@@ -1,0 +1,5 @@
+
+export default function() {
+  console.log("I'm the default");
+}
+export const named = {}