+++ test/fixtures/export-mixed/export-mixed-arrow.js
@@ -3,4 +3,4 @@ export default () => {
   return 1;
 }
 
-// export const two = 2;
+export const two = 2;