+++ test/fixtures/export-mixed/export-mixed-named-func.js
@@ -2,5 +2,6 @@
 export default function def() {
   console.log("I'm the default");
 }
+def.hey = 'hey'
 export { def } // TODO detect the self reference and don't extend
 export const named = {}